tests/wpt/web-platform-tests/web-bundle/subresource-loading/subresource-loading-relative-url-with-base.https.tentative.html

<!DOCTYPE html>
<title>Subresource loading using relative URLs in the 'resources' attribute with a base element</title>
<base href="../resources/wbn/static-element/">
<link rel="help" href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>

<body>
  <link
  rel="webbundle"
  href="../static-element.wbn"
  resources="resources/script.js"/>
  <script id="script" src="resources/script.js"></script>

  <script>
    const onLoadPromise = new Promise((resolve) => {
      window.addEventListener('load', resolve, false);
    });

    promise_test(async () => {
      await onLoadPromise;

      assert_equals(resources_script_result, 'loaded from webbundle');
    }, "A subresource script.js should be loaded from WebBundle using the relative URL and a base element.");
  </script>
</body>
